requirements.txt
pyspark==4.1.1
python-dotenv
dbt-dremio
dagster
dagster-cloud
dagster-dbt
dagster-webserver
boto3